+++ src-tauri/src/main.rs
@@ -6,30 +6,20 @@
 #[cfg(target_os = "windows")]
 mod windows;
 
+mod farmer;
 mod menu;
 mod node;
 
 use anyhow::Result;
-use log::{debug, error, info};
+use log::{debug, error};
 use serde::Serialize;
 use std::path::PathBuf;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
-use std::time::Duration;
-use subspace_core_primitives::{PublicKey, PIECE_SIZE};
-use subspace_farmer::{
-    Commitments, FarmerData, Farming, Identity, ObjectMappings, Plot, Plotting, RpcClient, WsRpc,
-};
-use subspace_solving::SubspaceCodec;
+use std::sync::atomic::Ordering;
 use tauri::SystemTrayEvent;
 use tauri::{
     api::{self},
     Env, Manager, RunEvent,
 };
-use tokio::runtime::Handle;
-
-static PLOTTED_PIECES: AtomicUsize = AtomicUsize::new(0);
-const BEST_BLOCK_NUMBER_CHECK_INTERVAL: Duration = Duration::from_secs(5);
 
 #[derive(Serialize)]
 struct DiskStats {
@@ -39,27 +29,41 @@ struct DiskStats {
 
 #[tauri::command]
 fn plot_progress_tracker() -> usize {
-    PLOTTED_PIECES.load(Ordering::Relaxed)
+    farmer::PLOTTED_PIECES.load(Ordering::Relaxed)
 }
 
 #[tauri::command]
-async fn farming(path: String, reward_address: String, plot_size: u64) -> bool {
-    if let Ok(address) = parse_reward_address(&reward_address) {
-        farm(path.into(), "ws://127.0.0.1:9944", Some(address), plot_size)
-            .await
-            .unwrap();
-        true
-    } else {
-        // reward address could not be parsed, and farmer did not start
-        false
+async fn farming(path: String, reward_address: String, plot_size: u64) {
+    match reward_address.len() {
+        0 => {
+            farmer::farm(path.into(), "ws://127.0.0.1:9944", None, plot_size)
+                .await
+                .unwrap();
+        }
+        _ => {
+            if let Ok(address) = farmer::parse_reward_address(&reward_address) {
+                farmer::farm(path.into(), "ws://127.0.0.1:9944", Some(address), plot_size)
+                    .await
+                    .unwrap();
+            } else {
+                error!("Reward address could not be parsed!");
+            }
+        }
     }
 }
 
 #[tauri::command]
 async fn start_node(path: String) -> String {
     let node_name = node::generate_node_name();
-    init_node(path.into(), node_name.clone()).await.unwrap();
+    node::init_node(path.into(), node_name.clone())
+        .await
+        .unwrap();
     node_name
+}
+
+#[tauri::command]
+fn identity_create(path: String) -> String {
+    farmer::create_identity(path.into()).unwrap()
 }
 
 #[tauri::command]
@@ -157,135 +161,4 @@ async fn main() -> Result<()> {
     });
 
     Ok(())
-}
-
-async fn init_node(base_directory: PathBuf, node_name: String) -> Result<()> {
-    let chain_spec =
-        sc_service::GenericChainSpec::<subspace_runtime::GenesisConfig>::from_json_bytes(
-            include_bytes!("../chain-spec.json").as_ref(),
-        )
-        .map_err(anyhow::Error::msg)?;
-
-    let full_client_fut = tokio::task::spawn_blocking(move || {
-        Handle::current().block_on(node::create_full_client(
-            chain_spec,
-            base_directory,
-            node_name,
-        ))
-    });
-    let mut full_client = full_client_fut.await??;
-
-    // TODO: Make this interruptable if needed
-    tokio::spawn(async move {
-        if let Err(error) = full_client.task_manager.future().await {
-            error!("Task manager exited with error: {error}");
-        } else {
-            error!("Task manager exited without error");
-        }
-    });
-
-    Ok(())
-}
-
-/// Start farming by using plot in specified path and connecting to WebSocket server at specified
-/// address.
-async fn farm(
-    base_directory: PathBuf,
-    node_rpc_url: &str,
-    reward_address: Option<PublicKey>,
-    plot_size: u64,
-) -> Result<()> {
-    let identity = Identity::open_or_create(&base_directory)?;
-    let address = identity.public_key().to_bytes().into();
-
-    let reward_address = reward_address.unwrap_or_else(|| identity.public_key().to_bytes().into());
-
-    info!("Connecting to node at {}", node_rpc_url);
-    let client = WsRpc::new(node_rpc_url).await?;
-
-    let farmer_metadata = client
-        .farmer_metadata()
-        .await
-        .map_err(|error| anyhow::Error::msg(error.to_string()))?;
-
-    // TODO: This doesn't account for the fact that node can
-    // have a completely different history to what farmer expects
-    info!("Opening plot");
-    let plot_fut = tokio::task::spawn_blocking({
-        let base_directory = base_directory.clone();
-        let plot_size = plot_size / PIECE_SIZE as u64;
-
-        // TODO: Piece count should account for database overhead of various additional databases
-        move || Plot::open_or_create(&base_directory, address, Some(plot_size))
-    });
-    let plot = plot_fut.await.unwrap()?;
-
-    // Keep track of the plotting for Desktop App
-    plot.on_progress_change(Arc::new(|plotted_pieces| {
-        PLOTTED_PIECES.fetch_add(
-            plotted_pieces.plotted_piece_count / PIECE_SIZE,
-            Ordering::SeqCst,
-        );
-        debug!(
-            "Plotted pieces so far: {}",
-            PLOTTED_PIECES.load(Ordering::Relaxed)
-        );
-    }))
-    .detach();
-
-    info!("Opening commitments");
-    let commitments_fut = tokio::task::spawn_blocking({
-        let path = base_directory.join("commitments");
-
-        move || Commitments::new(path)
-    });
-    let commitments = commitments_fut.await.unwrap()?;
-
-    info!("Opening object mapping");
-    let object_mappings = tokio::task::spawn_blocking({
-        let base_directory = base_directory.clone();
-
-        move || ObjectMappings::open_or_create(&base_directory)
-    })
-    .await??;
-
-    let subspace_codec = SubspaceCodec::new(identity.public_key());
-
-    // start the farming task
-    let farming_instance = Farming::start(
-        plot.clone(),
-        commitments.clone(),
-        client.clone(),
-        identity.clone(),
-        reward_address,
-    );
-
-    let farmer_data = FarmerData::new(plot, commitments, object_mappings, farmer_metadata);
-
-    // start the background plotting
-    let plotting_instance = Plotting::start(
-        farmer_data,
-        client,
-        subspace_codec,
-        BEST_BLOCK_NUMBER_CHECK_INTERVAL,
-    );
-
-    // wait for the farming and plotting in the background
-    tokio::spawn(async {
-        tokio::select! {
-            res = plotting_instance.wait() => if let Err(error) = res {
-                error!("Plotting created the error: {error}");
-            },
-            res = farming_instance.wait() => if let Err(error) = res {
-                error!("Farming created the error: {error}");
-            },
-        }
-    });
-
-    Ok(())
-}
-
-fn parse_reward_address(s: &str) -> Result<PublicKey, sp_core::crypto::PublicError> {
-    s.parse::<sp_core::sr25519::Public>()
-        .map(|key| PublicKey::from(key.0))
 }
